removed circular import between gui.classgui and gui.masks
