+++ suite2p/gui/classgui.py
@@ -4,11 +4,9 @@ import shutil
 import numpy as np
 from PyQt5 import QtGui
 
-from . import masks, io
+from . import masks
 from .. import classification
 
-
-#def make_buttons(parent)
 
 def make_buttons(parent,b0):
     # ----- CLASSIFIER BUTTONS -------
@@ -171,12 +169,6 @@ def add_to(parent):
         msg = QtGui.QMessageBox.information(parent,'Classifier saved and loaded',
                                             'Current dataset added to classifier, and cell probabilities computed and in GUI')
 
-def apply(parent):
-    classval = float(parent.probedit.text())
-    iscell = parent.probcell > classval
-    masks.flip_for_class(parent, iscell)
-    parent.update_plot()
-    io.save_iscell(parent)
 
 def save(parent, train_stats, train_iscell, keys):
     name = QtGui.QFileDialog.getSaveFileName(parent,'Classifier name (*.npy)')
@@ -204,7 +196,7 @@ def save_list(parent):
 def activate(parent, inactive):
     if inactive:
         parent.probcell = parent.model.predict_proba(parent.stat)
-    masks.class_masks(parent)
+    class_masks(parent)
     parent.update_plot()
 
 def disable(parent):
@@ -314,3 +306,16 @@ class ListChooser(QtGui.QDialog):
 
     def exit_list(self):
         self.accept()
+
+
+def class_masks(parent):
+    c = 6
+    istat = parent.probcell
+    parent.colors['colorbar'][c] = [istat.min(), (istat.max()-istat.min())/2, istat.max()]
+    istat = istat - istat.min()
+    istat = istat / istat.max()
+    col = masks.istat_transform(istat, parent.ops_plot['colormap'])
+    parent.colors['cols'][c] = col
+    parent.colors['istat'][c] = istat.flatten()
+
+    masks.rgb_masks(parent, col, c)
